Move is_complete() function from the redcap-det ETL to the common redcap module
This is nice to have more generically.
