+++ lib/id3c/cli/redcap.py
@@ -4,6 +4,7 @@ Minimal library for interacting with REDCap's web API.
 import logging
 import re
 import requests
+from enum import Enum
 from operator import itemgetter
 from typing import Any, Dict, List
 
@@ -117,3 +118,38 @@ class Project:
         response.raise_for_status()
 
         return response.json()
+
+
+class InstrumentStatus(Enum):
+    """
+    Numeric and string codes used by REDCap for instrument status.
+    """
+    Incomplete = 0
+    Unverified = 1
+    Complete = 2
+
+
+def is_complete(instrument: str, data: dict) -> bool:
+    """
+    Test if the named *instrument* is marked complete in the given *data*.
+
+    The *data* may be a DET notification or a record.
+
+    >>> is_complete("test", {"test_complete": "Complete"})
+    True
+    >>> is_complete("test", {"test_complete": 2})
+    True
+    >>> is_complete("test", {"test_complete": "2"})
+    True
+    >>> is_complete("test", {"test_complete": "Incomplete"})
+    False
+    >>> is_complete("test", {})
+    Traceback (most recent call last):
+        ...
+    KeyError: 'test_complete'
+    """
+    return data[f"{instrument}_complete"] in {
+        InstrumentStatus.Complete.name,
+        InstrumentStatus.Complete.value,
+        str(InstrumentStatus.Complete.value)
+    }
